removed multi-form view, added recipes.recipe-create view, refactored migrations and store recipe data logic

+++ database/migrations/2025_02_17_000012_ingredient_recipe.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('recipe_id')->constrained()->cascadeOnDelete();
             $table->foreignId('ingredient_id')->constrained()->cascadeOnDelete();
-            $table->decimal('quantity', 8, 2);
+            $table->integer('quantity');
             $table->foreignId('unit_id')->constrained();
             $table->timestamps();
         });
